requirements.txt
asgiref==3.6.0
awsebcli==3.20.6
backports.zoneinfo==0.2.1
boto3==1.26.129
botocore==1.29.129
cement==2.8.2
certifi==2022.12.7
charset-normalizer==2.0.12
colorama==0.4.3
Django==4.2
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-environ==0.10.0
django-js-asset==2.0.0
django-storages==1.13.2
django-taggit==3.1.0
djangorestframework==3.14.0
idna==3.4
jmespath==1.0.1
pathspec==0.10.1
Pillow==9.5.0
psycopg2==2.9.6
pypiwin32==223
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.6.1
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.4
termcolor==1.1.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.1.9
